Sample app shows master / detail in tablet landscape.
Some Flow API tweaks in support of that, and just general cleanup.
In particular, most reset methods take direction param. Still
not sure what to do with resetBackTo though, which emits either
BACKWARD or FORWARD.

The app will not win any graphic design awards. But The actual code to
split the screens is pretty sweet.